ref(test): further split models into struct files
