+++ src/api/blocks.rs
@@ -1,7 +1,9 @@
 use crate::http::client::Client;
 use std::borrow::Borrow;
 
-use crate::api::models::{Block, Transaction};
+use crate::api::models_new::block::Block;
+use crate::api::models_new::transaction::Transaction;
+
 use crate::api::Result;
 
 pub struct Blocks {
