+++ src/api/delegates.rs
@@ -2,7 +2,8 @@ use crate::http::client::Client;
 use std::borrow::Borrow;
 use std::collections::HashMap;
 
-use crate::api::models::{Balances, Block, Delegate, Wallet};
+use crate::api::models::{Balances, Delegate, Wallet};
+use crate::api::models_new::block::Block;
 use crate::api::Result;
 
 pub struct Delegates {
