+++ src/api/models_new/asset.rs
@@ -1,0 +1,38 @@
+use serde::Deserialize;
+use serde::Serialize;
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Asset {
+    #[serde(skip)]
+    None,
+    Signature {
+        #[serde(rename = "publicKey")]
+        public_key: String,
+    },
+    Delegate {
+        username: String,
+    },
+    Votes(Vec<String>),
+    #[serde(rename = "multisignature")]
+    MultiSignatureRegistration {
+        min: u8,
+        keysgroup: Vec<String>,
+        lifetime: u8,
+    },
+}
+
+impl Asset {
+    pub fn is_none(&self) -> bool {
+        match *self {
+            Asset::None => true,
+            _ => false,
+        }
+    }
+}
+
+impl Default for Asset {
+    fn default() -> Self {
+        Asset::None
+    }
+}
