+++ src/api/models_new/mod.rs
@@ -1,2 +1,6 @@
+pub mod asset;
+pub mod block;
 pub mod fee;
 pub mod node;
+pub mod timestamp;
+pub mod transaction;
