+++ tests/api/blocks_test.rs
@@ -1,9 +1,10 @@
 use serde_json::to_string_pretty;
+use serde_json::{from_str, Value};
 
-use crate::common::{mock_client, mock_http_request};
+use crate::common::{assert_transaction_data, mock_client, mock_http_request};
 
 #[test]
-fn test_all_blocks() {
+fn test_blocks_all() {
     let (_mock, body) = mock_http_request("blocks");
     {
         let mut client = mock_client();
@@ -19,20 +20,29 @@ fn test_show() {
     {
         let mut client = mock_client();
         let response = client.blocks.show("dummy").unwrap();
-        let actual = to_string_pretty(&response).unwrap();
-        assert_eq!(actual, body);
+        let expected: Value = from_str(&body).unwrap();
+
+        //TODO: add more tests, with type comparission. old actual/body are not relevant any more
+        assert_eq!(
+            response.data.payload.hash,
+            expected["data"]["payload"]["hash"].as_str().unwrap()
+        );
     }
 }
 
 #[test]
-fn test_transactions() {
+fn test_block_transactions() {
     let (_mock, body) = mock_http_request("blocks/dummy/transactions");
     {
         let mut client = mock_client();
         let response = client.blocks.transactions("dummy").unwrap();
+        let expected: Value = from_str(&body).unwrap();
 
-        let actual = to_string_pretty(&response).unwrap();
-        assert_eq!(actual, body);
+        for i in 0..=2 {
+            let rest_trx = response.data[i].clone();
+            let desered_trx = expected["data"][i].clone();
+            assert_transaction_data(rest_trx, &desered_trx);
+        }
     }
 }
