+++ tests/api/votes_test.rs
@@ -1,7 +1,8 @@
 use serde_json::from_str;
 
 use crate::common::*;
-use arkecosystem_client::api::models::{Asset, Transaction};
+use arkecosystem_client::api::models_new::asset::Asset;
+use arkecosystem_client::api::models_new::transaction::Transaction;
 use serde_json::Value;
 
 #[test]
