+++ tests/fixtures/blocks-dummy.json
@@ -1,30 +1,31 @@
 {
-  "data": {
-    "id": "17683763999369339115",
-    "version": 0,
-    "height": 972692,
-    "previous": "5213475885855819756",
-    "forged": {
-      "reward": 200000000,
-      "fee": 0,
-      "total": 200000000,
-      "amount": 0
-    },
-    "payload": {
-      "hash": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
-      "length": 0
-    },
-    "generator": {
-      "username": "gloom",
-      "address": "DS3tZr252ob1tSSeUD4R337qGqrWYjz4mV",
-      "publicKey": "0293f0d39a6fa3457fc4adac139ab58fea8532699e5cdc78faa345d4fdcf942b32"
-    },
-    "signature": "3045022100d79b7096a1511eacb8e714e1f370201ce480bd56e1c809cd6f4e75348144d7190220391f32aaaf32b7e12c84988f12aee1efadab21a850b0ab63ffa7b99991cb0735",
-    "transactions": 0,
-    "timestamp": {
-      "epoch": 54335560,
-      "unix": 1544436760,
-      "human": "2018-12-10T10:12:40.000Z"
+    "data": {
+        "id": "994e048f3933c730757afee36303914f7f0739eb5f5d497a9fc7b733abc86602",
+        "version": 0,
+        "height": 4168948,
+        "previous": "a36685cb1ac0d6974e771db68711acf625f0cc3a1fce89c3d5ed154fcffea450",
+        "forged": {
+            "reward": "200000000",
+            "fee": "0",
+            "total": "200000000",
+            "amount": "0"
+        },
+        "payload": {
+            "hash": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+            "length": 0
+        },
+        "generator": {
+            "username": "alexxio",
+            "address": "DEz1Mr4uJ7NaiufwKEj28atCRPTmsUqh9t",
+            "publicKey": "0215789ac26155b7a338708f595b97c453e08918d0630c896cbd31d83fe2ad1c33"
+        },
+        "signature": "304402203c4657d20c6c6f12bce8824d68d223a56596c806824e15c4c5249648bac43920022060971f1a4484d98effac8f8710ea7aae3bf69ac9629d2f4b15a3d130d0a0a161",
+        "confirmations": 0,
+        "transactions": 0,
+        "timestamp": {
+            "epoch": 83722464,
+            "unix": 1573823664,
+            "human": "2019-11-15T13:14:24.000Z"
+        }
     }
-  }
 }
